Adding Encoder To/FromUtfX methods
