+++ src/System.Text.Primitives/System/Text/Encoding/Encoders.Utf32.cs
@@ -8,38 +8,38 @@ namespace System.Text.Encoders
 {
     public static class Utf32
     {
-        #region Utf-8 to Utf-32 conversion
+        #region UTF-8 Conversions
 
-        public static TransformationStatus ComputeEncodedBytesFromUtf8(ReadOnlySpan<byte> source, out int bytesNeeded)
+        public static TransformationStatus FromUtf8Length(ReadOnlySpan<byte> source, out int bytesNeeded)
+            => Utf8.ToUtf32Length(source, out bytesNeeded);
+
+        public static TransformationStatus FromUtf8(ReadOnlySpan<byte> source, Span<byte> destination, out int bytesConsumed, out int bytesWritten)
+            => Utf8.ToUtf32(source, destination, out bytesConsumed, out bytesWritten);
+
+        public static TransformationStatus ToUtf8Length(ReadOnlySpan<byte> source, out int bytesNeeded)
         {
             bytesNeeded = 0;
 
-            int index = 0;
-            int length = source.Length;
-            ref byte src = ref source.DangerousGetPinnableReference();
+            ref uint utf32 = ref Unsafe.As<byte, uint>(ref source.DangerousGetPinnableReference());
+            int utf32Length = source.Length >> 2; // byte => uint count
 
-            while (index < length)
+            for (int i = 0; i < utf32Length; i++)
             {
-                int count = EncodingHelper.GetUtf8DecodedBytes(Unsafe.Add(ref src, index));
-                if (count == 0)
-                    goto InvalidData;
-                if (length - index < count)
-                    goto NeedMoreData;
+                uint codePoint = Unsafe.Add(ref utf32, i);
+                if (!EncodingHelper.IsSupportedCodePoint(codePoint))
+                    return TransformationStatus.InvalidData;
 
-                bytesNeeded += count;
+                bytesNeeded += EncodingHelper.GetUtf8EncodedBytes(codePoint);
             }
 
-            return index < length ? TransformationStatus.DestinationTooSmall : TransformationStatus.Done;
+            if (utf32Length << 2 != source.Length)
+                return TransformationStatus.NeedMoreSourceData;
 
-        InvalidData:
-            return TransformationStatus.InvalidData;
-
-        NeedMoreData:
-            return TransformationStatus.NeedMoreSourceData;
+            return TransformationStatus.Done;
         }
 
         /// <summary>
-        /// Converts a span containing a sequence of UTF-8 bytes into UTF-32 bytes.
+        /// Converts a span containing a sequence of UTF-32 bytes into UTF-8 bytes.
         ///
         /// This method will consume as many of the input bytes as possible.
         ///
@@ -47,150 +47,179 @@ namespace System.Text.Encoders
         /// equal to the length of the <paramref name="source"/> and <paramref name="bytesWritten"/> will equal the total number of bytes written to
         /// the <paramref name="destination"/>.
         /// </summary>
-        /// <param name="source">A span containing a sequence of UTF-8 bytes.</param>
-        /// <param name="destination">A span to write the UTF-32 bytes into.</param>
+        /// <param name="source">A span containing a sequence of UTF-32 bytes.</param>
+        /// <param name="destination">A span to write the UTF-8 bytes into.</param>
         /// <param name="bytesConsumed">On exit, contains the number of bytes that were consumed from the <paramref name="source"/>.</param>
         /// <param name="bytesWritten">On exit, contains the number of bytes written to <paramref name="destination"/></param>
         /// <returns>A <see cref="TransformationStatus"/> value representing the state of the conversion.</returns>
-        public static TransformationStatus ConvertFromUtf8(ReadOnlySpan<byte> source, Span<byte> destination, out int bytesConsumed, out int bytesWritten)
+        public static TransformationStatus ToUtf8(ReadOnlySpan<byte> source, Span<byte> destination, out int bytesConsumed, out int bytesWritten)
         {
             bytesConsumed = 0;
             bytesWritten = 0;
 
-            int srcLength = source.Length;
-            int dstLength = destination.Length;
             ref byte src = ref source.DangerousGetPinnableReference();
+            int srcLength = source.Length;
+
             ref byte dst = ref destination.DangerousGetPinnableReference();
+            int dstLength = destination.Length;
 
-            while (bytesConsumed < srcLength && bytesWritten < dstLength)
+            while (srcLength - bytesConsumed >= sizeof(uint))
             {
-                uint codePoint = Unsafe.Add(ref src, bytesConsumed);
+                uint codePoint = Unsafe.As<byte, uint>(ref Unsafe.Add(ref src, bytesConsumed));
+                if (!EncodingHelper.IsSupportedCodePoint(codePoint))
+                    return TransformationStatus.InvalidData;
 
-                int byteCount = EncodingHelper.GetUtf8DecodedBytes((byte)codePoint);
-                if (byteCount == 0)
-                    goto InvalidData;
-                if (srcLength - bytesConsumed < byteCount)
-                    goto NeedMoreData;
+                int bytesNeeded = EncodingHelper.GetUtf8EncodedBytes(codePoint);
+                if (dstLength - bytesWritten < bytesNeeded)
+                    return TransformationStatus.DestinationTooSmall;
 
-                if (byteCount > 1)
-                    codePoint &= (byte)(0x7F >> byteCount);
-
-                for (var i = 1; i < byteCount; i++)
+                switch (bytesNeeded)
                 {
-                    ref byte next = ref Unsafe.Add(ref src, bytesConsumed + i);
-                    if ((next & EncodingHelper.b1100_0000U) != EncodingHelper.b1000_0000U)
-                        goto InvalidData;
+                    case 1:
+                        Unsafe.Add(ref dst, bytesWritten) = (byte)(EncodingHelper.b0111_1111U & codePoint);
+                        break;
 
-                    codePoint = (codePoint << 6) | (uint)(EncodingHelper.b0011_1111U & next);
+                    case 2:
+                        Unsafe.Add(ref dst, bytesWritten) = (byte)(((codePoint >> 6) & EncodingHelper.b0001_1111U) | EncodingHelper.b1100_0000U);
+                        Unsafe.Add(ref dst, bytesWritten + 1) = (byte)((codePoint & EncodingHelper.b0011_1111U) | EncodingHelper.b1000_0000U);
+                        break;
+
+                    case 3:
+                        Unsafe.Add(ref dst, bytesWritten) = (byte)(((codePoint >> 12) & EncodingHelper.b0000_1111U) | EncodingHelper.b1110_0000U);
+                        Unsafe.Add(ref dst, bytesWritten + 1) = (byte)(((codePoint >> 6) & EncodingHelper.b0011_1111U) | EncodingHelper.b1000_0000U);
+                        Unsafe.Add(ref dst, bytesWritten + 2) = (byte)((codePoint & EncodingHelper.b0011_1111U) | EncodingHelper.b1000_0000U);
+                        break;
+
+                    case 4:
+                        Unsafe.Add(ref dst, bytesWritten) = (byte)(((codePoint >> 18) & EncodingHelper.b0000_0111U) | EncodingHelper.b1111_0000U);
+                        Unsafe.Add(ref dst, bytesWritten + 1) = (byte)(((codePoint >> 12) & EncodingHelper.b0011_1111U) | EncodingHelper.b1000_0000U);
+                        Unsafe.Add(ref dst, bytesWritten + 2) = (byte)(((codePoint >> 6) & EncodingHelper.b0011_1111U) | EncodingHelper.b1000_0000U);
+                        Unsafe.Add(ref dst, bytesWritten + 3) = (byte)((codePoint & EncodingHelper.b0011_1111U) | EncodingHelper.b1000_0000U);
+                        break;
+
+                    default:
+                        return TransformationStatus.InvalidData;
                 }
 
-                Unsafe.As<byte, uint>(ref Unsafe.Add(ref dst, bytesWritten)) = codePoint;
-                bytesWritten += 4;
-                bytesConsumed += byteCount;
+                bytesConsumed += 4;
+                bytesWritten += bytesNeeded;
             }
 
-            return bytesConsumed < srcLength ? TransformationStatus.DestinationTooSmall : TransformationStatus.Done;
-
-        InvalidData:
-            return TransformationStatus.InvalidData;
-
-        NeedMoreData:
-            return TransformationStatus.NeedMoreSourceData;
+            return bytesConsumed < srcLength ? TransformationStatus.NeedMoreSourceData : TransformationStatus.Done;
         }
+
+        #endregion UTF-8 Conversions
+
+        #region UTF-16 Conversions
+
+        public static TransformationStatus FromUtf16Length(ReadOnlySpan<byte> source, out int bytesNeeded)
+            => Utf16.ToUtf32Length(source, out bytesNeeded);
+
+        public static TransformationStatus FromUtf16(ReadOnlySpan<byte> source, Span<byte> destination, out int bytesConsumed, out int bytesWritten)
+            => Utf16.ToUtf32(source, destination, out bytesConsumed, out bytesWritten);
+
+        /// <summary>
+        /// Calculates the byte count needed to encode the UTF-16 bytes from the specified UTF-32 sequence.
+        ///
+        /// This method will consume as many of the input bytes as possible.
+        /// </summary>
+        /// <param name="source">A span containing a sequence of UTF-32 bytes.</param>
+        /// <param name="bytesNeeded">On exit, contains the number of bytes required for encoding from the <paramref name="source"/>.</param>
+        /// <returns>A <see cref="TransformationStatus"/> value representing the expected state of the conversion.</returns>
+        public static TransformationStatus ToUtf16Length(ReadOnlySpan<byte> source, out int bytesNeeded)
+        {
+            int index = 0;
+            int length = source.Length;
+            ref byte src = ref source.DangerousGetPinnableReference();
+
+            bytesNeeded = 0;
+
+            while (length - index >= 4)
+            {
+                ref uint codePoint = ref Unsafe.As<byte, uint>(ref Unsafe.Add(ref src, index));
+
+                if (!EncodingHelper.IsSupportedCodePoint(codePoint))
+                    return TransformationStatus.InvalidData;
+
+                bytesNeeded += EncodingHelper.IsBmp(codePoint) ? 2 : 4;
+                index += 4;
+            }
+
+            return index < length ? TransformationStatus.NeedMoreSourceData : TransformationStatus.Done;
+        }
+
+        /// <summary>
+        /// Converts a span containing a sequence of UTF-32 bytes into UTF-16 bytes.
+        ///
+        /// This method will consume as many of the input bytes as possible.
+        ///
+        /// On successful exit, the entire input was consumed and encoded successfully. In this case, <paramref name="bytesConsumed"/> will be
+        /// equal to the length of the <paramref name="source"/> and <paramref name="bytesWritten"/> will equal the total number of bytes written to
+        /// the <paramref name="destination"/>.
+        /// </summary>
+        /// <param name="source">A span containing a sequence of UTF-32 bytes.</param>
+        /// <param name="destination">A span to write the UTF-16 bytes into.</param>
+        /// <param name="bytesConsumed">On exit, contains the number of bytes that were consumed from the <paramref name="source"/>.</param>
+        /// <param name="bytesWritten">On exit, contains the number of bytes written to <paramref name="destination"/></param>
+        /// <returns>A <see cref="TransformationStatus"/> value representing the state of the conversion.</returns>
+        public static TransformationStatus ToUtf16(ReadOnlySpan<byte> source, Span<byte> destination, out int bytesConsumed, out int bytesWritten)
+        {
+            ref byte src = ref source.DangerousGetPinnableReference();
+            ref byte dst = ref destination.DangerousGetPinnableReference();
+            int srcLength = source.Length;
+            int dstLength = destination.Length;
+
+            bytesConsumed = 0;
+            bytesWritten = 0;
+
+            while (srcLength - bytesConsumed >= sizeof(uint))
+            {
+                ref uint codePoint = ref Unsafe.As<byte, uint>(ref Unsafe.Add(ref src, bytesConsumed));
+
+                if (!EncodingHelper.IsSupportedCodePoint(codePoint))
+                    return TransformationStatus.InvalidData;
+
+                int written = EncodingHelper.IsBmp(codePoint) ? 2 : 4;
+                if (dstLength - bytesWritten < written)
+                    return TransformationStatus.DestinationTooSmall;
+
+                unchecked
+                {
+                    if (written == 2)
+                        Unsafe.As<byte, char>(ref Unsafe.Add(ref dst, bytesWritten)) = (char)codePoint;
+                    else
+                    {
+                        Unsafe.As<byte, char>(ref Unsafe.Add(ref dst, bytesWritten)) = (char)(((codePoint - 0x010000u) >> 10) + EncodingHelper.HighSurrogateStart);
+                        Unsafe.As<byte, char>(ref Unsafe.Add(ref dst, bytesWritten + 2)) = (char)((codePoint & 0x3FF) + EncodingHelper.LowSurrogateStart);
+                    }
+                }
+
+                bytesWritten += written;
+                bytesConsumed += 4;
+            }
+
+            return bytesConsumed < srcLength ? TransformationStatus.NeedMoreSourceData : TransformationStatus.Done;
+        }
+
+        #endregion UTF-16 Conversions
+
+        #region Utf-8 to Utf-32 conversion
+
+        public static TransformationStatus ComputeEncodedBytesFromUtf8(ReadOnlySpan<byte> source, out int bytesNeeded)
+            => Utf8.ToUtf32Length(source, out bytesNeeded);
+
+        public static TransformationStatus ConvertFromUtf8(ReadOnlySpan<byte> source, Span<byte> destination, out int bytesConsumed, out int bytesWritten)
+            => Utf8.ToUtf32(source, destination, out bytesConsumed, out bytesWritten);
 
         #endregion Utf-8 to Utf-32 conversion
 
         #region Utf-16 to Utf-32 conversion
 
         public static TransformationStatus ComputeEncodedBytesFromUtf16(ReadOnlySpan<byte> source, out int bytesNeeded)
-        {
-            bytesNeeded = 0;
+            => Utf16.ToUtf32Length(source, out bytesNeeded);
 
-            ref byte src = ref source.DangerousGetPinnableReference();
-            int srcLength = source.Length;
-            int srcIndex = 0;
-
-            while (srcLength - srcIndex >= sizeof(char))
-            {
-                uint codePoint = Unsafe.As<byte, char>(ref Unsafe.Add(ref src, srcIndex));
-                if (EncodingHelper.IsSurrogate(codePoint))
-                {
-                    if (!EncodingHelper.IsHighSurrogate(codePoint))
-                        return TransformationStatus.InvalidData;
-
-                    if (srcLength - srcIndex < sizeof(char) * 2)
-                        return TransformationStatus.NeedMoreSourceData;
-
-                    uint lowSurrogate = Unsafe.As<byte, char>(ref Unsafe.Add(ref src, srcIndex + 2));
-                    if (!EncodingHelper.IsLowSurrogate(lowSurrogate))
-                        return TransformationStatus.InvalidData;
-
-                    srcIndex += 2;
-                }
-
-                srcIndex += 2;
-                bytesNeeded += 4;
-            }
-
-            return srcIndex < srcLength ? TransformationStatus.NeedMoreSourceData : TransformationStatus.Done;
-        }
-
-        /// <summary>
-        /// Converts a span containing a sequence of UTF-16 bytes into UTF-32 bytes.
-        ///
-        /// This method will consume as many of the input bytes as possible.
-        ///
-        /// On successful exit, the entire input was consumed and encoded successfully. In this case, <paramref name="bytesConsumed"/> will be
-        /// equal to the length of the <paramref name="source"/> and <paramref name="bytesWritten"/> will equal the total number of bytes written to
-        /// the <paramref name="destination"/>.
-        /// </summary>
-        /// <param name="source">A span containing a sequence of UTF-16 bytes.</param>
-        /// <param name="destination">A span to write the UTF-32 bytes into.</param>
-        /// <param name="bytesConsumed">On exit, contains the number of bytes that were consumed from the <paramref name="source"/>.</param>
-        /// <param name="bytesWritten">On exit, contains the number of bytes written to <paramref name="destination"/></param>
-        /// <returns>A <see cref="TransformationStatus"/> value representing the state of the conversion.</returns>
         public static TransformationStatus ConvertFromUtf16(ReadOnlySpan<byte> source, Span<byte> destination, out int bytesConsumed, out int bytesWritten)
-        {
-            bytesConsumed = 0;
-            bytesWritten = 0;
-
-            ref byte src = ref source.DangerousGetPinnableReference();
-            int srcLength = source.Length;
-
-            ref byte dst = ref destination.DangerousGetPinnableReference();
-            int dstLength = destination.Length;
-
-            while (srcLength - bytesConsumed >= sizeof(char))
-            {
-                if (dstLength - bytesWritten < sizeof(uint))
-                    return TransformationStatus.DestinationTooSmall;
-
-                uint codePoint = Unsafe.As<byte, char>(ref Unsafe.Add(ref src, bytesConsumed));
-                if (EncodingHelper.IsSurrogate(codePoint))
-                {
-                    if (!EncodingHelper.IsHighSurrogate(codePoint))
-                        return TransformationStatus.InvalidData;
-
-                    if (srcLength - bytesConsumed < sizeof(char) * 2)
-                        return TransformationStatus.NeedMoreSourceData;
-
-                    uint lowSurrogate = Unsafe.As<byte, char>(ref Unsafe.Add(ref src, bytesConsumed + 2));
-                    if (!EncodingHelper.IsLowSurrogate(lowSurrogate))
-                        return TransformationStatus.InvalidData;
-
-                    codePoint -= EncodingHelper.HighSurrogateStart;
-                    lowSurrogate -= EncodingHelper.LowSurrogateStart;
-                    codePoint = ((codePoint << 10) | lowSurrogate) + 0x010000u;
-                    bytesConsumed += 2;
-                }
-
-                Unsafe.As<byte, uint>(ref Unsafe.Add(ref dst, bytesWritten)) = codePoint;
-                bytesConsumed += 2;
-                bytesWritten += 4;
-            }
-
-            return bytesConsumed < srcLength ? TransformationStatus.NeedMoreSourceData : TransformationStatus.Done;
-        }
+            => Utf16.ToUtf32(source, destination, out bytesConsumed, out bytesWritten);
 
         #endregion Utf-16 to Utf-32 conversion
     }
